requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# NLP & ML
spacy==3.8.4
sentence-transformers==2.7.0
transformers==4.40.2
torch==2.2.2

# Vector DB
faiss-cpu==1.8.0

# Resume Parsing
pdfplumber==0.11.0
python-docx==1.1.0
PyPDF2==3.0.1

# Excel
openpyxl==3.1.2
xlsxwriter==3.2.0
pandas==2.2.2

# Email / Microsoft Graph
msal==1.28.0
requests==2.31.0
httpx==0.27.0

# Async
aiofiles==23.2.1

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.3.0
ujson==5.9.0

# Dev / testing
pytest==8.2.0
pytest-asyncio==0.23.6